generateRPointShapes {easyGgplot2}	R Documentation

Generate a plot of point shapes which R knows about.

Description

Generate point shapes available in R

Usage

generateRPointShapes()

Author(s)

Alboukadel Kassambara <alboukadel.kassambara@gmail.com>

References

http://www.sthda.com

See Also

generateRLineTypes

Examples

#The code to generate point shapes available in R
generateRPointShapes()

[Package easyGgplot2 version 1.0.0.9000 Index]