showCols {easyGgplot2}	R Documentation

Generate a plot of color names which R knows about.

Description

Generate R color names.

Usage

showCols(cl = colors(), bg = "grey", cex = 0.75, rot = 30)

Arguments

cl
A vector of colors to plot. Default value is the list of color names returned by R color() function.

bg
Background color of the plot

cex
The size of the text.

rot
The rotation angle of the text

Value

a ggplot

Author(s)

Alboukadel Kassambara <alboukadel.kassambara@gmail.com>

References

http://www.sthda.com

See Also

generateRLineTypes, generateRPointShapes

Examples

col=colors() #list of color names
showCols(col[1:100]) #plot the first 100 colors

[Package easyGgplot2 version 1.0.0.9000 Index]